started new lab verification hook
